# TM8/tm8-game-mobile-release-0.6.0/input_folder/dev/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/notifications/token/refresh": {
            "post": {
                "operationId": "NotificationController_refreshNotificationToken",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationRefreshTokenDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK response"
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications/message": {
            "post": {
                "operationId": "NotificationController_createMessageNotification",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMessageNotificationDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK response"
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications/call": {
            "post": {
                "operationId": "NotificationController_createCallNotification",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCallNotificationDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK response"
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications/unread/count": {
            "get": {
                "operationId": "NotificationController_checkUnreadNotifications",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnreadNotificationsResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications": {
            "get": {
                "operationId": "NotificationController_listUsers",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "minimum": 1,
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Page size",
                        "schema": {
                            "minimum": 1,
                            "maximum": 50,
                            "default": 10,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationPaginatedResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications/settings": {
            "patch": {
                "operationId": "NotificationController_updateNotificationSettings",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNotificationSettingsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "204": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications/settings/options": {
            "get": {
                "operationId": "NotificationController_getNotificationSettingsOptions",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NotificationSettingsOptionResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications/{notificationId}": {
            "delete": {
                "operationId": "NotificationController_deleteNotification",
                "parameters": [
                    {
                        "name": "notificationId",
                        "required": true,
                        "in": "path",
                        "description": "Notification ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "204": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            },
            "get": {
                "operationId": "NotificationController_getNotificationDetail",
                "parameters": [
                    {
                        "name": "notificationId",
                        "required": true,
                        "in": "path",
                        "description": "Notification ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
        "/api/v1/notifications/{notificationId}/read": {
            "patch": {
                "operationId": "NotificationController_markNotificationAsRead",
                "parameters": [
                    {
                        "name": "notificationId",
                        "required": true,
                        "in": "path",
                        "description": "Notification ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "204": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Notification"
                ]
            }
        },
...
    },
    "info": {
        "title": "TM8 API",
        "description": "TM8 API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "PaginationMetaResponse": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "number",
                        "description": "Page number"
                    },
                    "limit": {
                        "type": "number",
                        "description": "Page size"
                    },
                    "itemCount": {
                        "type": "number",
                        "description": "Total items without pagination"
                    },
                    "pageCount": {
                        "type": "number",
                        "description": "Total pages with current page size"
                    },
                    "hasPreviousPage": {
                        "type": "boolean",
                        "description": "Flag indicating if there is a previous page"
                    },
                    "hasNextPage": {
                        "type": "boolean",
                        "description": "Flag indicating if there is a next page"
                    }
                },
                "required": [
                    "page",
                    "limit",
                    "itemCount",
                    "pageCount",
                    "hasPreviousPage",
                    "hasNextPage"
                ]
            },
...
            "TGMDExceptionResponse": {
                "type": "object",
                "properties": {
                    "origin": {
                        "enum": [
                            "HTTP",
                            "Mongo",
                            "TGMD",
                            "Unknown"
                        ],
                        "type": "string"
                    },
                    "status": {
                        "type": "number",
                        "description": "HTTP-compatible status code"
                    },
                    "code": {
                        "type": "string"
                    },
                    "exception": {
                        "type": "string",
                        "description": "Exception name"
                    },
                    "detail": {
                        "type": "string",
                        "description": "Human-readable error message"
                    }
                },
                "required": [
                    "origin",
                    "status",
                    "code",
                    "exception",
                    "detail"
                ]
            },
...
            "NotificationRefreshTokenDto": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "required": [
                    "token"
                ]
            },
            "CreateMessageNotificationDto": {
                "type": "object",
                "properties": {
                    "senderUsername": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "recipientId": {
                        "type": "string"
                    },
                    "redirectScreen": {
                        "type": "string"
                    }
                },
                "required": [
                    "senderUsername",
                    "message",
                    "recipientId",
                    "redirectScreen"
                ]
            },
            "CreateCallNotificationDto": {
                "type": "object",
                "properties": {
                    "callerUsername": {
                        "type": "string"
                    },
                    "recipientId": {
                        "type": "string"
                    },
                    "redirectScreen": {
                        "type": "string"
                    }
                },
                "required": [
                    "callerUsername",
                    "recipientId",
                    "redirectScreen"
                ]
            },
            "UnreadNotificationsResponse": {
                "type": "object",
                "properties": {
                    "unread": {
                        "type": "boolean"
                    },
                    "count": {
                        "type": "number"
                    }
                },
                "required": [
                    "unread"
                ]
            },
            "NotificationDataResponse": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "redirectScreen": {
                        "type": "string"
                    },
                    "isRead": {
                        "type": "boolean"
                    },
                    "notificationType": {
                        "enum": [
                            "game-update",
                            "new-features",
                            "system-maintenance",
                            "exclusive-offers",
                            "community-news",
                            "other",
                            "call",
                            "message",
                            "friend-request",
                            "friend-added",
                            "match",
                            "reminder",
                            "ban",
                            "suspend"
                        ],
                        "type": "string"
                    }
                },
                "required": [
                    "title",
                    "description",
                    "redirectScreen",
                    "isRead",
                    "notificationType"
                ]
            },
            "NotificationResponse": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "user": {
                        "type": "string"
                    },
                    "since": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "until": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/NotificationDataResponse"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "_id",
                    "user",
                    "since",
                    "until",
                    "data",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "NotificationPaginatedResponse": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NotificationResponse"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMetaResponse"
                    }
                },
                "required": [
                    "items",
                    "meta"
                ]
            },
            "UpdateNotificationSettingsDto": {
                "type": "object",
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    },
                    "match": {
                        "type": "boolean"
                    },
                    "friendAdded": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "boolean"
                    },
                    "reminders": {
                        "type": "boolean"
                    },
                    "news": {
                        "type": "boolean"
                    }
                }
            },
            "NotificationSettingsOptionResponse": {
                "type": "object",
                "properties": {
                    "keyValue": {
                        "type": "string"
                    },
                    "displayValue": {
                        "type": "string"
                    }
                },
                "required": [
                    "keyValue",
                    "displayValue"
                ]
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}